Increase readability and rework on file load/save mechanism
