+++ src/fastga_he/gui/power_train_builder/__init__.py
@@ -1,0 +1,5 @@
+# This file is part of FAST-OAD_CS23-HE : A framework for rapid Overall Aircraft Design of Hybrid
+# Electric Aircraft.
+# Copyright (C) 2026 ISAE-SUPAERO
+
+from .power_train_builder_launcher import PowertrainBuilderLauncher
